refactor emails to own package
