+++ apps/web/package.json
@@ -70,7 +70,8 @@
     "tailwind-merge": "^2.3.0",
     "tailwindcss-animate": "^1.0.7",
     "use-debounce": "^10.0.1",
-    "zod": "^3.23.8"
+    "zod": "^3.23.8",
+    "emails": "workspace:*"
   },
   "devDependencies": {
     "@cloudflare/next-on-pages": "^1.13.10",
