+++ packages/emails/components/RegistrationConfirmation.tsx
@@ -16,10 +16,15 @@ import {
 } from "@react-email/components";
 import type * as React from "react";
 import c, { emailsConfig } from "config";
+import { DefaultFooter, DefaultHeader } from "./shared";
 
 const baseUrl = process.env.NEXT_PUBLIC_BASE_URL!;
 
-export default function AccountConnected({ firstName }: { firstName: string }) {
+export default function RegistrationConfirmation({
+	firstName,
+}: {
+	firstName: string;
+}) {
 	return (
 		<Html>
 			<Head />
@@ -40,22 +45,16 @@ export default function AccountConnected({ firstName }: { firstName: string }) {
 					},
 				}}
 			>
-				<Preview>{`Welcome back to ${c.universityName} ${c.clubName}. Thanks for connecting your account. We are
-									glad to have you back with us. Things have
-									changed since you were last here. Lots of
-									updates and enhancements that we hope you
-									will enjoy.`}</Preview>
+				<Preview>{`Welcome to ${c.universityName} ${c.clubName}. Thanks for registering with us and we are
+									happy to have you as one of our members. As
+									you might know, ACM is always dedicated to
+									our members and we want to help you get
+									familiar with us and our membership portal.`}</Preview>
 				<Body className="bg-offwhite font-sans text-base">
-					<Img
-						src={emailsConfig.publicLogoLink}
-						width="100"
-						height="100"
-						alt="Logo"
-						className="mx-auto my-20"
-					/>
+					<DefaultHeader />
 					<Container className="p-45 bg-white">
 						<Heading className="my-0 text-center leading-8">
-							{`Welcome back to ${c.universityName} ${c.clubName}`}
+							{`Welcome to ${c.universityName} ${c.clubName}`}
 						</Heading>
 
 						<Section className="">
@@ -64,11 +63,11 @@ export default function AccountConnected({ firstName }: { firstName: string }) {
 									{`Hi ${firstName},`}
 								</Text>
 								<Text className="text-base">
-									Thanks for connecting your account. We are
-									glad to have you back with us. Things have
-									changed since you were last here. Lots of
-									updates and enhancements that we hope you
-									will enjoy.
+									Thanks for registering with us and we are
+									happy to have you as one of our members. As
+									you might know, ACM is always dedicated to
+									our members and we want to help you get
+									familiar with us and our membership portal.
 								</Text>
 								<Text className="mt-4 text-base">
 									Here's how to get started:
@@ -88,8 +87,10 @@ export default function AccountConnected({ firstName }: { firstName: string }) {
 							</li>
 							<li className="mb-20">
 								<strong>Check out our upcoming events! </strong>
-								{`${c.clubName} is always hosting cool events that give you an opportunity to come learn, have snacks, and meet new people so be sure to check out the upcoming `}
-								<Link href={`${baseUrl}/events`}>events.</Link>
+								{`${c.clubName} is always hosting cool events that give you an opportunity to come learn, have snacks, and meet new people so be sure to `}
+								<Link href={`${baseUrl}/events`}>
+									check out what is happening soon.
+								</Link>
 							</li>
 							<li className="mb-20">
 								<strong>
@@ -112,26 +113,23 @@ export default function AccountConnected({ firstName }: { firstName: string }) {
 							</Button>
 						</Section>
 
-						<div className="mt-45 flex w-full flex-wrap items-center justify-center gap-x-4">
-							{emailsConfig.footerLinks.map((link) => (
-								<div key={link.name}>
-									<Link
-										className="font-bold text-black underline"
-										href={link.href}
-									>
-										{link.name}
-									</Link>{" "}
-									<span className="text-green-500">→</span>
-								</div>
-							))}
-						</div>
+						<Row className="mt-45 max-w-xs gap-x-10">
+													{emailsConfig.footerLinks.map((link) => (
+														<Column key={link.name} className="text-center">
+																<Link
+																	className="font-bold text-black underline "
+																	href={link.href}
+																>
+																	{link.name}
+																</Link>{" "}
+																<span className="text-green-500">
+																	→
+																</span>
+														</Column>
+													))}
+												</Row>
 					</Container>
-
-					<Container className="mt-20">
-						<Text className="mb-45 text-center text-gray-400">
-							{`${emailsConfig.rightsReservedString}`}
-						</Text>
-					</Container>
+					<DefaultFooter/>
 				</Body>
 			</Tailwind>
 		</Html>
