+++ packages/emails/components/index.ts
@@ -1,0 +1,2 @@
+export {default as AccountConnected} from "./AccountConnected"
+export {default as RegistrationConfirmation} from "./RegistrationConfirmation"
