+++ packages/emails/index.ts
@@ -1,0 +1,2 @@
+export * as components from "./components/index";
+export * as utils from "./utils";
