+++ packages/emails/package.json
@@ -1,0 +1,25 @@
+{
+  "name": "emails",
+  "version": "1.0.0",
+  "description": "",
+ "main": "./index.tsx",
+ "types": "./index.tsx",
+  "scripts": {
+    "dev": "email dev --dir ./components --port 3001"
+  },
+  "keywords": [],
+  "author": "",
+  "license": "ISC",
+  "devDependencies": {
+    "@types/node": "^20.11.24",
+    "react-email": "4.0.4"
+  },
+  "dependencies": {
+    "@react-email/components": "0.0.36",
+    "@types/react": "^18",
+    "@types/react-dom": "^18",
+    "config": "workspace:*",
+    "react": "^18",
+    "react-dom": "^18"
+  }
+}
